pyproject.toml
[project]
name = "curso-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.11",
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "openai>=1.100.2",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
